requirements.txt
dgl==0.3
h5py==2.9.0
ipdb==0.12
matplotlib==3.1.0
numpy==1.16.4
Pillow==6.0.0
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.2.1
six==1.12.0
tensorboard==1.14.0
tensorboardX==1.8
tensorflow==1.14.0
tensorflow-estimator==1.14.0
torch==1.1.0
torchvision==0.3.0
tqdm==4.32.2